refactoring of voting view to avoid code duplications
The `conference.views.voting` callback function has been refactored and modularised into multiple functions in order to reduce as much as possible
code duplications in p3 views.

The OptionForm Class has been moved to `conference.forms`, as it is extended by a new form in `p3`.
Moreover, a date format change is also committed to fix a bug about TalkSubmission form.
